requirements.txt
requests
pandas
ta
psycopg2-binary
yfinance
pyyaml
